dev: Migrate to WSL and refactor files
- moved wrappers and templatetags folders from util_tools in backend to better split the project into backend-frontend
- migrated to WSL and changed backend base.html for Linux filesystem
- renamed the Django settings folder from leafsee_simple to configs to better match the purpose of the folder
- the rest of the files were commits by git itself, don't know why
