+++ .gitignore
@@ -107,6 +107,8 @@ celerybeat.pid
 env/
 venv/
 ENV/
+.env.bak/
+.venv.bak/
 env.bak/
 venv.bak/
