+++ backend/apps/templates/base/base.html
@@ -27,12 +27,12 @@
                     class="clip-polygon-octagon float-left flex size-10 items-center justify-center hover:bg-green-l-hover active:bg-green-l-onclick"
                     onclick="toggleElement('base-sidebar')"
                 >
-                    {% load_svg 'svg\burger_menu_button.svg' class='w-8' %}
+                    {% load_svg 'svg/burger_menu_button.svg' class='w-8' %}
                 </button>
                 <button
                     class="float-left ml-5 flex h-9 w-max items-center bg-green-l px-4 clip-polygon-steep-2"
                 >
-                    {% load_svg 'svg\leaf_icon.svg' class='w-7 float-left' %}
+                    {% load_svg 'svg/leaf_icon.svg' class='w-7 float-left' %}
                     <p
                         class="float-left ml-2 text-center font-jockeyone text-3xl"
                     >
@@ -44,7 +44,7 @@
                 <button
                     class="float-left h-full bg-green-l pl-5 pr-3 text-center hover:bg-green-l-hover active:bg-green-l-onclick"
                 >
-                    {% load_svg 'svg\magnifier.svg' class='w-5' %}
+                    {% load_svg 'svg/magnifier.svg' class='w-5' %}
                 </button>
                 <input
                     type="text"
@@ -56,16 +56,16 @@
             {% if user.is_authenticated %}
             <div class="flex h-full w-max items-center justify-between gap-4">
                 <a href="{% url 'video-list' %}">
-                    {% load_svg 'svg\video_upload.svg' class='w-8' %}
+                    {% load_svg 'svg/video_upload.svg' class='w-8' %}
                 </a>
                 <button onclick="toggleElement('base-notifications')">
-                    {% load_svg 'svg\notification_bell.svg' class='w-6' %}
+                    {% load_svg 'svg/notification_bell.svg' class='w-6' %}
                 </button>
                 <button
                     class="clip-polygon-octagon flex size-8 items-center justify-center bg-white"
                     onclick="toggleElement('base-account-submenu')"
                 >
-                    {% load_svg 'svg\user_icon.svg' class='w-7' %}
+                    {% load_svg 'svg/user_icon.svg' class='w-7' %}
                 </button>
             </div>
             {% else %}
@@ -123,7 +123,7 @@
                             <div
                                 class="clip-polygon-octagon flex size-8 items-center justify-center bg-white"
                             >
-                                {% load_svg 'svg\user_icon.svg' class='w-7' %}
+                                {% load_svg 'svg/user_icon.svg' class='w-7' %}
                             </div>
                             <p class="font-play text-xl">Название канала</p>
                         </button>
