+++ backend/configs/tests.py
@@ -1,0 +1,5 @@
+"""
+Module for run tests from apps
+"""
+
+from apps.videos.tests import *  # noqa: F401 F403
