--- backend/leafsee_simple/example.env
@@ -1,12 +1,0 @@
-# Django secret key
-SECRET_KEY="secret"
-
-# If TRUE activate django debug mod
-DEBUG=<TRUE or FALSE>
-
-# Database parapets in django-environment library formate, for MySQL, PostgreSQL or SQLite
-DATABASE_URL=postgres://<user>:<password>@<ip>:<port>/<database name>
-    -OR-
-DATABASE_URL=mysql://<user>:<password>@<ip>:<port>/<database name>
-    -OR-
-DATABASE_URL=sqlite:///<path/to/sqlite.db>
