--- backend/leafsee_simple/tests.py
@@ -1,5 +1,0 @@
-"""
-Module for run tests from apps
-"""
-
-from apps.videos.tests import *  # noqa: F401 F403
