--- backend/leafsee_simple/urls.py
@@ -1,28 +1,0 @@
-"""
-# URL configuration for leafsee_simple project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/5.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-
-from django.contrib import admin
-from django.conf import settings
-from django.urls import path, include
-from django.conf.urls.static import static
-
-
-urlpatterns = [
-    path("admin/", admin.site.urls),
-    path("auth/", include("apps.authentication.urls")),
-    path("", include("apps.videos.urls")),
-] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
